Parametric bootstrap standard errors and confidence intervals

Simulates n_boot series of the same length from the fitted model, refits each one, and summarizes the resulting sampling distribution of the estimates. This is the recommended alternative to the Hessian standard errors of gs_fit() when the asymptotic normal approximation is unreliable: near a boundary (stable tail index near 2, GEV shape near \(\pm 0.5\), APARCH asymmetry near \(\pm 1\)) or, for GEV innovations, because the support depends on the shape parameter and the usual regularity conditions do not hold (Efron, 1979).

Usage

gs_bootstrap(fit, n_boot = 199, level = 0.95, seed = NULL, ...)

Arguments

fit

A gs_fit object.

n_boot

Number of bootstrap replications.

level

Confidence level of the percentile intervals.

seed

Optional seed for reproducibility.

...

Passed to gs_fit() (for example algorithm).

Value

A list with se (named vector of bootstrap standard errors), ci (a matrix with lower and upper percentile limits), replicates (the n_boot by number-of-parameters matrix of estimates, failed fits as NA rows), n_fail and level.

Details

Each replication is refit with hessian = FALSE; replications whose fit fails are dropped and counted in n_fail.

References

Efron, B. (1979). Bootstrap methods: another look at the jackknife. The Annals of Statistics, 7(1), 1-26.

See also

Examples

# \donttest{
spec <- gs_spec(garch = c(1, 1), dist = gs_gev())
model <- gs_model(spec, mu = 0, omega = 0.1, alpha = 0.1, beta = 0.8,
                  dist_par = c(xi = 0.2))
x <- gs_sim(model, n = 800, seed = 1)$y
#> Warning: Persistence 1.2018 >= 1: the simulated scale has no stationary level.
fit <- gs_fit(x, spec)
#> Warning: Parameter(s) at a bound, standard errors set to NA: omega.
#> Warning: Persistence 1.2030 >= 1: E(sigma_t^delta) is not finite and long horizon scale forecasts diverge.
gs_bootstrap(fit, n_boot = 99, seed = 1)
#> $se
#>           mu        omega       alpha1        beta1           xi 
#> 9.629267e+14 8.945189e+32 2.003281e-01 1.417758e-01 5.338741e-02 
#> 
#> $ci
#>                lower        upper
#> mu     -3.256099e+03 1.416524e+08
#> omega   8.172852e+08 4.925955e+17
#> alpha1  1.059683e-01 8.775871e-01
#> beta1   2.643308e-01 7.776756e-01
#> xi      1.650510e-03 2.029478e-01
#> 
#> $replicates
#>                   mu        omega     alpha1     beta1          xi
#>   [1,]  1.910895e+04 4.454121e+09 0.12921380 0.7470162  0.15730763
#>   [2,]  3.968717e+04 1.606493e+10 0.31867547 0.5595214  0.11400970
#>   [3,]  4.565100e+04 2.311300e+09 0.11928060 0.7571281  0.11651337
#>   [4,]  6.795916e+04 6.277707e+10 0.12035473 0.7525663  0.15971762
#>   [5,]  2.173389e+04 1.710068e+09 0.21170436 0.6544288  0.12723261
#>   [6,]  1.633861e+04 2.234893e+09 0.12559962 0.7446972  0.10887304
#>   [7,]  5.557614e+04 9.682379e+09 0.16608113 0.6994208  0.20019826
#>   [8,]  1.582815e+04 1.238784e+09 0.10001692 0.7784279  0.12937774
#>   [9,]  2.512217e+07 1.269632e+16 0.32512231 0.5341575  0.16835999
#>  [10,]  2.620808e+04 2.725627e+09 0.14435865 0.7200446  0.14580335
#>  [11,]  5.948234e+04 9.544737e+09 0.21431643 0.6438704  0.17777589
#>  [12,]  1.887009e+04 2.268649e+09 0.10842610 0.7737282  0.12367309
#>  [13,] -7.174058e+03 6.995268e+08 0.19948901 0.6664803  0.18017545
#>  [14,]  3.607812e+05 2.503617e+12 0.61997319 0.4114957  0.13903204
#>  [15,]  1.716541e+04 1.560185e+09 0.12789183 0.7319537  0.15742854
#>  [16,]  1.172997e+04 2.724943e+09 0.10967387 0.7687667  0.13379693
#>  [17,]  1.371696e+04 2.026241e+09 0.20594817 0.6582797  0.16634475
#>  [18,]  3.354965e+06 2.887340e+14 0.14072806 0.7314423  0.18382824
#>  [19,]  3.729330e+04 9.733617e+09 0.16641904 0.6785242  0.16196940
#>  [20,]  3.441415e+04 8.120981e+09 0.17863097 0.6672421  0.11627247
#>  [21,]  1.838140e+04 1.270238e+09 0.09386963 0.7771880  0.17182117
#>  [22,]  1.410224e+04 6.964803e+08 0.11395081 0.7557754  0.20331414
#>  [23,]  2.399236e+06 5.932486e+13 0.38693843 0.5268247  0.14750137
#>  [24,]  1.152901e+06 1.109651e+14 0.77336942 0.2770351  0.01678662
#>  [25,]  1.439428e+04 2.699092e+09 0.16091222 0.7050097  0.16080127
#>  [26,]  6.888654e+06 8.516843e+14 0.16438130 0.7124437  0.16437794
#>  [27,]  8.188631e+06 4.500888e+14 0.19422467 0.6655657  0.14333448
#>  [28,]  4.001880e+04 2.715851e+09 0.15904237 0.7368409 -0.01073358
#>  [29,]  7.382315e+04 5.409920e+10 0.24766393 0.6092417  0.14305047
#>  [30,]  1.584451e+04 1.997984e+09 0.14276831 0.7166565  0.20110674
#>  [31,]  1.135366e+09 8.813796e+17 0.14264831 0.7362500  0.18884587
#>  [32,]  9.580999e+15 8.900351e+33 0.88859240 0.2536637  0.04225127
#>  [33,]  1.938547e+06 9.381391e+12 0.21446214 0.6643477  0.18083780
#>  [34,]  2.266209e+04 2.188733e+09 0.20099618 0.6567388  0.09050407
#>  [35,]  1.204896e+04 1.318739e+09 0.14880992 0.7139578  0.12673292
#>  [36,]  8.344472e+03 2.753759e+10 0.13538685 0.7446089  0.14717514
#>  [37,]  1.279477e+04 1.210188e+09 0.11931907 0.7570153  0.15745429
#>  [38,]  1.048973e+04 2.291426e+09 0.19353992 0.6614963  0.15034802
#>  [39,]  1.267658e+05 1.172161e+11 0.17915554 0.6741609  0.19946876
#>  [40,]  4.025099e+05 4.244145e+12 0.54154632 0.3875924  0.07393424
#>  [41,]  6.411581e+06 1.741508e+16 0.73838529 0.3654426  0.05541473
#>  [42,]  2.582312e+04 5.886538e+09 0.15846284 0.7201221  0.13339831
#>  [43,]  2.686139e+04 4.840088e+09 0.14782067 0.7145828  0.19493868
#>  [44,]  4.180586e+04 2.091514e+09 0.29853684 0.6660416 -0.07169419
#>  [45,]  2.795308e+04 1.456106e+10 0.24656620 0.6157605  0.15240149
#>  [46,]  5.532177e+03 1.442187e+09 0.12681589 0.7418297  0.15147255
#>  [47,]  2.464065e+04 2.404893e+09 0.19768761 0.6648593  0.15538523
#>  [48,]  6.425938e+05 1.064819e+13 0.21905155 0.6463626  0.16243610
#>  [49,] -2.624516e+03 1.464514e+09 0.15132354 0.7302888  0.15553471
#>  [50,]  1.252836e+05 6.358514e+10 0.32613761 0.5624293  0.11510685
#>  [51,]  2.031789e+03 7.766362e+08 0.12326537 0.7521134  0.13471633
#>  [52,]  2.209214e+05 3.661535e+11 0.32495256 0.5369730  0.12637642
#>  [53,]  5.352893e+04 5.327997e+09 0.17816138 0.7448369 -0.07549479
#>  [54,]  1.384405e+04 2.412396e+09 0.12710212 0.7350631  0.18671883
#>  [55,]  2.289545e+08 1.739990e+18 0.20490372 0.6553906  0.18328306
#>  [56,]  1.630136e+02 1.036287e+09 0.11121181 0.7595456  0.12315113
#>  [57,]  6.373430e+03 1.674887e+09 0.17191770 0.6768560  0.18515918
#>  [58,]  6.450799e+05 4.032418e+13 0.66332559 0.3785611  0.10760483
#>  [59,]  9.659205e+03 2.888296e+09 0.12862714 0.7380920  0.15663762
#>  [60,]  2.823312e+03 1.477228e+09 0.11190443 0.7526854  0.17579870
#>  [61,]  7.044658e+03 2.800306e+09 0.18110671 0.6770953  0.19135408
#>  [62,]  2.012580e+04 1.486354e+09 0.21257309 0.6620710  0.09832316
#>  [63,]  1.929965e+06 2.678624e+14 0.89969847 0.2966831  0.09659329
#>  [64,]  2.926497e+04 2.327857e+09 0.16946058 0.6886311  0.13299883
#>  [65,]  8.908759e+04 4.144861e+10 0.40005620 0.4552885  0.20253066
#>  [66,]  4.231155e+04 9.177243e+09 0.11118297 0.7674452  0.02535450
#>  [67,]  1.760668e+04 1.914513e+09 0.19355471 0.6728441  0.17955660
#>  [68,]  2.893990e+05 9.148420e+11 0.77955902 0.1359482  0.24581615
#>  [69,]  5.110209e+04 1.425256e+10 0.27213708 0.6017193  0.15697723
#>  [70,]  9.493094e+03 1.076414e+09 0.11138153 0.7780745  0.12327541
#>  [71,]  6.735833e+03 1.798156e+09 0.13081072 0.7306317  0.14163637
#>  [72,]  7.005962e+04 7.700949e+10 0.22440638 0.6277542  0.17720535
#>  [73,]  3.005709e+04 3.926009e+09 0.17410113 0.6924656  0.19354630
#>  [74,]  1.355310e+04 1.488901e+09 0.13541907 0.7299775  0.12488910
#>  [75,]  2.459988e+04 1.610106e+09 0.11714202 0.7600783  0.09942900
#>  [76,]  1.226166e+03 8.669673e+08 0.12332794 0.7484498  0.15249882
#>  [77,]  1.557911e+04 2.302032e+09 0.14807060 0.7175812  0.14553316
#>  [78,] -3.772850e+03 1.101194e+09 0.12959333 0.7322785  0.18987066
#>  [79,]  1.808665e+06 1.996383e+14 0.87473686 0.2712625  0.14168713
#>  [80,]  1.122056e+04 1.276840e+09 0.11860290 0.7566792  0.15055881
#>  [81,]  2.575401e+05 9.749672e+11 0.52226336 0.4548419  0.03639363
#>  [82,]  1.017296e+07 1.476628e+15 0.44772479 0.4762713  0.09633729
#>  [83,]  3.494978e+07 9.320869e+15 0.37100549 0.5386241  0.13680109
#>  [84,]  4.550129e+05 2.768851e+12 0.64562184 0.3665937  0.20328918
#>  [85,]  1.137020e+04 1.195166e+09 0.14045269 0.7101439  0.17391274
#>  [86,]  2.396932e+04 5.042908e+09 0.18983064 0.6499037  0.18626991
#>  [87,]  2.927241e+06 7.424865e+15 0.87991911 0.2586594  0.07354772
#>  [88,]  1.044940e+04 2.088003e+10 0.13065069 0.7532559  0.17276662
#>  [89,]  1.698873e+04 7.048136e+09 0.13920535 0.7252969  0.14085947
#>  [90,]  2.696765e+04 6.094928e+09 0.24029199 0.6195099  0.15676103
#>  [91,]  1.209316e+04 2.967248e+09 0.14465414 0.7003330  0.16809195
#>  [92,]  2.373827e+05 2.973579e+11 0.24756383 0.6189805  0.14682806
#>  [93,]  7.346496e+03 1.117660e+09 0.12622181 0.7485896  0.14211377
#>  [94,]  9.024544e+04 1.322971e+11 0.10395741 0.7837885  0.18104406
#>  [95,]  2.354533e+04 1.146270e+09 0.14380268 0.7343891  0.16268000
#>  [96,]  1.716151e+04 1.645688e+09 0.11813824 0.7660597  0.13328189
#>  [97,]  4.799270e+04 6.028690e+09 0.12829884 0.7419660  0.14351498
#>  [98,]  5.279975e+03 6.741572e+09 0.27101606 0.5921556  0.20018864
#>  [99,] -6.635561e+03 9.971232e+09 0.20003738 0.6696668  0.17961071
#> 
#> $n_fail
#> [1] 0
#> 
#> $level
#> [1] 0.95
#> 
# }